Change the sidebar with the simplest one and make the layout content
we use flex to make content separate to be two, navbar or sidebar and the main content. And at the beginning we want to make the side bar sticky but it can't work because it's inside flex parents, so the solution is to make flex-item (in this case the sidebar which want to be sticky) to be flex-start.